bpaf_derive for custom_usage
