+++ bpaf_derive/src/field_tests.rs
@@ -816,6 +816,18 @@ fn hide_usage() {
 }
 
 #[test]
+fn custom_usage() {
+    let input: NamedField = parse_quote! {
+        #[bpaf(custom_usage(usage()))]
+        field: u32
+    };
+    let output = quote! {
+        ::bpaf::long("field").argument::<u32>("ARG").custom_usage(usage())
+    };
+    assert_eq!(input.to_token_stream().to_string(), output.to_string());
+}
+
+#[test]
 fn argument_with_manual_parse() {
     let input: NamedField = parse_quote! {
         #[bpaf(argument::<String>("N"), parse(twice_the_num))]
